Move local recipe search to new search route
