+++ openapi_paths.yaml
@@ -25,7 +25,7 @@ paths:
           name: q
           schema:
             type: string
-  /recipes/find:
+  /search:
     get:
       summary: Search local recipes
       parameters:
